+++ src/amateras/_version.py
@@ -1,5 +1,5 @@
 # coding: utf-8
 # file generated by setuptools_scm
 # don't change, don't track in version control
-__version__ = version = '0.1.dev21+g7cee689.d20220803'
-__version_tuple__ = version_tuple = (0, 1, 'dev21', 'g7cee689.d20220803')
+__version__ = version = '0.1.dev22+ge47fbe9.d20220816'
+__version_tuple__ = version_tuple = (0, 1, 'dev22', 'ge47fbe9.d20220816')
